Defined IExtensionState
IExtensionState will be the the extensions local state

It will consist of two parts, a session state and a shared state.
The session state will store information only usefull to the current
session, like caches, and other session relevant data. Data that should
not be stored on the database should be housed in the session state.

The shared state will be reflected on the database and updated on the
database when there are changed localy, and updated localy when there
are changed on the database. Session independent information should be
stored in the shared state so it won't be lost if the session stops.