Removed dbContext, mongoDbManager and baseDbManager, added graphql
Removed the code for accessing and modifying the database directly in
favour of using a new graphql api that will be unified between the bot
and the frontend.

Added new useGraphQL function and created graphql client.
Added new useJwtToken function for generating valid jwt access tokens
for the graphql backend